backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
python-dotenv
fastkml
shapely
geojson
beautifulsoup4
lxml
httpx